fix(fileutil): only remove writePath on error for atomic copy path
In the non-atomic path writePath==dst, so the error defer was deleting
the destination file on failures like Sync even though it may have
contained valid data. Only remove writePath when writePath != dst
(i.e., the temp file used in the atomic path).

Closes k8senv-yq8

+++ internal/fileutil/copy.go
@@ -67,7 +67,7 @@ func CopyFile(src, dst string, opts *CopyFileOptions) (retErr error) {
 		return err
 	}
 	defer func() {
-		if retErr != nil {
+		if retErr != nil && writePath != dst {
 			_ = os.Remove(writePath)
 		}
 	}()
